+++ core/relay/adaptor/anthropic/openai.go
@@ -415,7 +415,7 @@ func Response2OpenAI(meta *meta.Meta, claudeResponse *Response) *relaymodel.Text
 		},
 	}
 	if fullTextResponse.Usage.PromptTokens == 0 {
-		fullTextResponse.Usage.PromptTokens = meta.RequestUsage.InputTokens
+		fullTextResponse.Usage.PromptTokens = int64(meta.RequestUsage.InputTokens)
 	}
 	fullTextResponse.Usage.TotalTokens = fullTextResponse.Usage.PromptTokens + fullTextResponse.Usage.CompletionTokens
 	return &fullTextResponse
@@ -470,8 +470,8 @@ func OpenAIStreamHandler(m *meta.Meta, c *gin.Context, resp *http.Response) (*mo
 			}
 			usage.Add(response.Usage)
 			if usage.PromptTokens == 0 {
-				usage.PromptTokens = m.RequestUsage.InputTokens
-				usage.TotalTokens += m.RequestUsage.InputTokens
+				usage.PromptTokens = int64(m.RequestUsage.InputTokens)
+				usage.TotalTokens += int64(m.RequestUsage.InputTokens)
 			}
 			response.Usage = usage
 			responseText.Reset()
@@ -493,9 +493,9 @@ func OpenAIStreamHandler(m *meta.Meta, c *gin.Context, resp *http.Response) (*mo
 
 	if usage == nil {
 		usage = &relaymodel.Usage{
-			PromptTokens:     m.RequestUsage.InputTokens,
+			PromptTokens:     int64(m.RequestUsage.InputTokens),
 			CompletionTokens: openai.CountTokenText(responseText.String(), m.OriginModel),
-			TotalTokens:      m.RequestUsage.InputTokens + openai.CountTokenText(responseText.String(), m.OriginModel),
+			TotalTokens:      int64(m.RequestUsage.InputTokens) + openai.CountTokenText(responseText.String(), m.OriginModel),
 		}
 		_ = render.ObjectData(c, &relaymodel.ChatCompletionsStreamResponse{
 			ID:      openai.ChatCompletionID(),
